requirements.txt
tflite_runtime
numpy
Pillow